backend/requirements.txt
# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
Flask-PyMongo==2.3.0
pymongo==4.6.0

# Authentication & Security
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2

# AI/ML Dependencies
torch==2.1.2
torchvision==0.16.2
diffusers==0.25.0
transformers==4.36.2
accelerate==0.25.0

# Image Processing
Pillow==10.1.0

# Audio Processing
openai-whisper==20231117

# Translation
googletrans==4.0.0rc1

# System Monitoring
psutil==5.9.6

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0
python-dotenv==1.0.0

# Additional Utilities
requests==2.31.0
python-multipart==0.0.6 